requirements.txt
psutil
requests
pystray
Pillow
pyinstaller
plyer